requirements.txt
numpy>=1.26,<3.0
pandas>=2.0,<3.0
scikit-learn>=1.3,<2.0
xgboost>=2.0,<3.0
imbalanced-learn>=0.11,<1.0
torch>=2.1,<3.0
streamlit>=1.28,<2.0
optuna>=3.4,<5.0
matplotlib>=3.7,<4.0
seaborn>=0.12,<1.0
joblib>=1.3,<2.0
pyyaml>=6.0,<7.0
pytest>=7.4,<9.0
pyarrow>=14.0
